analysis_bot/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.16
pydantic-settings==2.2.1
apscheduler==3.10.4
httpx==0.27.0
python-telegram-bot>=21.4
yfinance
beautifulsoup4
google-genai
python-dotenv
jinja2
python-multipart
tenacity
feedparser
python-docx
python-docx
aiofiles
natsort
pandas

